Adds Deucalion.Application. Initial support for application configuration in yaml.

+++ Deucalion.Network/Monitors/Options/TcpMonitorOptions.cs
@@ -5,6 +5,6 @@ namespace Deucalion.Network.Monitors.Options
     public class TcpMonitorOptions : PullMonitorOptions
     {
         public string Host { get; set; } = default!;
-        public ushort Port { get; set; } = default!;
+        public int Port { get; set; } = default!;
     }
 }
